code refactor

+++ app/src/main/java/com/nxb/solidprinciples/ocp/good/Rectangle.kt
@@ -1,10 +1,10 @@
 package com.nxb.solidprinciples.ocp.good
 
 class Rectangle : Shape {
-    var length = 0
-    var height = 0
+    var length = 0.0
+    var height = 0.0
 
-    override fun getArea(): Int {
+    override fun getArea(): Double {
         return (length * height)
     }
 }
